GUAC-1132: Socket service is now really tunnel service.
