+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/tunnel/MultiseatGuacamoleTunnelService.java
@@ -20,13 +20,16 @@
  * THE SOFTWARE.
  */
 
-package org.glyptodon.guacamole.auth.jdbc.socket;
+package org.glyptodon.guacamole.auth.jdbc.tunnel;
 
 import com.google.inject.Singleton;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import org.glyptodon.guacamole.GuacamoleClientTooManyException;
 import org.glyptodon.guacamole.auth.jdbc.user.AuthenticatedUser;
 import org.glyptodon.guacamole.auth.jdbc.connection.ModeledConnection;
 import org.glyptodon.guacamole.GuacamoleException;
@@ -35,31 +38,51 @@ import org.glyptodon.guacamole.auth.jdbc.connectiongroup.ModeledConnectionGroup;
 
 
 /**
- * GuacamoleSocketService implementation which allows only one user per
- * connection at any time, but does not disallow concurrent use of connection
- * groups. If a user attempts to use a connection group multiple times, they
- * will receive different underlying connections each time until the group is
- * exhausted.
+ * GuacamoleTunnelService implementation which restricts concurrency only on a
+ * per-user basis. Each connection or group may be used concurrently any number
+ * of times, but each concurrent use must be associated with a different user.
  *
  * @author Michael Jumper
  */
 @Singleton
-public class BalancedGuacamoleSocketService
-    extends AbstractGuacamoleSocketService {
+public class MultiseatGuacamoleTunnelService
+    extends AbstractGuacamoleTunnelService {
 
     /**
-     * The set of all active connection identifiers.
+     * The set of all active user/connection pairs.
      */
-    private final Set<String> activeConnections =
-            Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
+    private final Set<Seat> activeSeats =
+            Collections.newSetFromMap(new ConcurrentHashMap<Seat, Boolean>());
 
+    /**
+     * The set of all active user/connection group pairs.
+     */
+    private final Set<Seat> activeGroupSeats =
+            Collections.newSetFromMap(new ConcurrentHashMap<Seat, Boolean>());
+   
     @Override
     protected ModeledConnection acquire(AuthenticatedUser user,
             List<ModeledConnection> connections) throws GuacamoleException {
 
-        // Return the first unused connection
-        for (ModeledConnection connection : connections) {
-            if (activeConnections.add(connection.getIdentifier()))
+        String username = user.getUser().getIdentifier();
+
+        // Sort connections in ascending order of usage
+        ModeledConnection[] sortedConnections = connections.toArray(new ModeledConnection[connections.size()]);
+        Arrays.sort(sortedConnections, new Comparator<ModeledConnection>() {
+
+            @Override
+            public int compare(ModeledConnection a, ModeledConnection b) {
+
+                return getActiveConnections(a).size()
+                     - getActiveConnections(b).size();
+
+            }
+
+        });
+        
+        // Return the first unreserved connection
+        for (ModeledConnection connection : sortedConnections) {
+            if (activeSeats.add(new Seat(username, connection.getIdentifier())))
                 return connection;
         }
 
@@ -70,19 +93,24 @@ public class BalancedGuacamoleSocketService
 
     @Override
     protected void release(AuthenticatedUser user, ModeledConnection connection) {
-        activeConnections.remove(connection.getIdentifier());
+        activeSeats.remove(new Seat(user.getUser().getIdentifier(), connection.getIdentifier()));
     }
 
     @Override
     protected void acquire(AuthenticatedUser user,
             ModeledConnectionGroup connectionGroup) throws GuacamoleException {
-        // Do nothing
+
+        // Do not allow duplicate use of connection groups
+        Seat seat = new Seat(user.getUser().getIdentifier(), connectionGroup.getIdentifier());
+        if (!activeGroupSeats.add(seat))
+            throw new GuacamoleClientTooManyException("Cannot connect. Connection group already in use by this user.");
+
     }
 
     @Override
     protected void release(AuthenticatedUser user,
             ModeledConnectionGroup connectionGroup) {
-        // Do nothing
+        activeGroupSeats.remove(new Seat(user.getUser().getIdentifier(), connectionGroup.getIdentifier()));
     }
 
 }
