+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/tunnel/UnrestrictedGuacamoleTunnelService.java
@@ -20,80 +20,63 @@
  * THE SOFTWARE.
  */
 
-package org.glyptodon.guacamole.auth.jdbc.socket;
+package org.glyptodon.guacamole.auth.jdbc.tunnel;
 
 import com.google.inject.Singleton;
-import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import org.glyptodon.guacamole.GuacamoleClientTooManyException;
 import org.glyptodon.guacamole.auth.jdbc.user.AuthenticatedUser;
 import org.glyptodon.guacamole.auth.jdbc.connection.ModeledConnection;
 import org.glyptodon.guacamole.GuacamoleException;
-import org.glyptodon.guacamole.GuacamoleResourceConflictException;
 import org.glyptodon.guacamole.auth.jdbc.connectiongroup.ModeledConnectionGroup;
 
 
 /**
- * GuacamoleSocketService implementation which allows exactly one use
- * of any connection at any time. Concurrent usage of connections is not
- * allowed, and concurrent usage of connection groups is allowed only between
- * different users.
+ * GuacamoleTunnelService implementation which imposes no restrictions
+ * whatsoever on the number of concurrent or duplicate connections.
  *
  * @author Michael Jumper
  */
 @Singleton
-public class SingleSeatGuacamoleSocketService
-    extends AbstractGuacamoleSocketService {
+public class UnrestrictedGuacamoleTunnelService
+    extends AbstractGuacamoleTunnelService {
 
-    /**
-     * The set of all active connection identifiers.
-     */
-    private final Set<String> activeConnections =
-            Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
-
-    /**
-     * The set of all active user/connection group pairs.
-     */
-    private final Set<Seat> activeGroupSeats =
-            Collections.newSetFromMap(new ConcurrentHashMap<Seat, Boolean>());
-   
     @Override
     protected ModeledConnection acquire(AuthenticatedUser user,
             List<ModeledConnection> connections) throws GuacamoleException {
 
-        // Return the first unused connection
+        ModeledConnection chosen = null;
+        int lowestUsage = 0;
+
+        // Find connection with lowest usage
         for (ModeledConnection connection : connections) {
-            if (activeConnections.add(connection.getIdentifier()))
-                return connection;
+
+            int usage = getActiveConnections(connection).size();
+            if (chosen == null || usage < lowestUsage) {
+                chosen = connection;
+                lowestUsage = usage;
+            }
+            
         }
 
-        // Already in use
-        throw new GuacamoleResourceConflictException("Cannot connect. This connection is in use.");
-
+        return chosen;
+        
     }
 
     @Override
     protected void release(AuthenticatedUser user, ModeledConnection connection) {
-        activeConnections.remove(connection.getIdentifier());
+        // Do nothing
     }
 
     @Override
     protected void acquire(AuthenticatedUser user,
             ModeledConnectionGroup connectionGroup) throws GuacamoleException {
-
-        // Do not allow duplicate use of connection groups
-        Seat seat = new Seat(user.getUser().getIdentifier(), connectionGroup.getIdentifier());
-        if (!activeGroupSeats.add(seat))
-            throw new GuacamoleClientTooManyException("Cannot connect. Connection group already in use by this user.");
-
+        // Do nothing
     }
 
     @Override
     protected void release(AuthenticatedUser user,
             ModeledConnectionGroup connectionGroup) {
-        activeGroupSeats.remove(new Seat(user.getUser().getIdentifier(), connectionGroup.getIdentifier()));
+        // Do nothing
     }
 
 }
