+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/tunnel/package-info.java
@@ -21,7 +21,7 @@
  */
 
 /**
- * Classes related to obtaining/configuring Guacamole sockets, and restricting
- * access to those sockets.
+ * Classes related to obtaining/configuring Guacamole tunnels, and restricting
+ * access to those tunnels.
  */
-package org.glyptodon.guacamole.auth.jdbc.socket;
+package org.glyptodon.guacamole.auth.jdbc.tunnel;
